backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=3.15
gunicorn
psycopg2-binary
python-dotenv
dj-database-url
